% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_confint.R
\name{confint_vector}
\alias{confint_vector}
\title{Confidence interval calulation for sample}
\usage{
confint_vector(x, sigma = -1, alpha = 0.05)
}
\arguments{
\item{x}{a numeric vector}

\item{sigma}{the standard deviation.If you know the sd value, please set it to \code{sigma}. Default is -1, which means that you didn't know the sd value.}

\item{alpha}{significance level}
}
\value{
a data frame of basic statistics information
}
\description{
Confidence interval calulation for sample
}
\examples{
set.seed(2019);x <- sample(1:1000,500,replace = T)
confint_vector(x)
confint_vector(x,sigma = 2) # Not be used as usual.It's because we always don't know the value of standard deviation.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
