% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_convert.R
\name{convert2}
\alias{convert2}
\title{multiple annotation based on dataset common.annot.}
\usage{
convert2(vt, fromtype = "ENTREZID", totype = "SYMBOL", db = common.annot)
}
\arguments{
\item{vt}{genes you want to annotation}

\item{fromtype}{the raw type of the genes.}

\item{totype}{the annotated type of the genes.}

\item{db}{a dataset containing annotation information. In lucky package, it is often \code{common.annot} dataset for \code{hg38}.}
}
\description{
multiple annotation based on dataset common.annot.Its colnames include:"ENSEMBL","ENSEMBL.Versions","SYMBOL","GENENAME","ENTREZID","UNIGENE" and "gene_type".
}
