% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countListElement.R
\name{countListElement}
\alias{countListElement}
\title{countListElement}
\usage{
countListElement(list)
}
\arguments{
\item{list}{a list object}
}
\value{
Numeric. The sum of bottom non-list objects in the list.
}
\description{
Calculate the sum of bottom non-list objects in a list
}
\examples{
l <- list(a=list(a1=1, a2=2), b=list(b1=1, b2=2))
countListElement(l) # 4
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
