% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_CPM2FPKM.R
\name{CPM2FPKM}
\alias{CPM2FPKM}
\title{CPM to FPKM}
\usage{
CPM2FPKM(
  mt,
  geneid = "ensembl",
  genomic = c("hg38", "hg38_biomart", "hg19")[1]
)
}
\arguments{
\item{mt}{Read count matrix(row is gene, column is sample)}

\item{geneid}{Only support \code{ensembl} now}

\item{genomic}{Only support \code{hg38}, \code{hg19}, or \code{hg38_biomart} now}
}
\value{
FPKM-like expression matrix
}
\description{
RNA-Seq CPM to FPKM RNA-Seq Read Count to FPKM-like Gene length-normalized expression matrix
}
\details{
FPKM-like expression = Read Count × 10^9 / (Gene Length(bp) × Total Read Count)
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
