% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_enhanceConvert.R
\name{enhanceConvert}
\alias{enhanceConvert}
\title{enhanceConvert}
\usage{
enhanceConvert(
  vt,
  fromtype = "SYMBOL",
  totype = "ENSEMBL",
  db = common.annot,
  enhanceList = NULL
)
}
\arguments{
\item{vt}{genes you want to annotation}

\item{fromtype}{the raw type of the genes.}

\item{totype}{the annotated type of the genes.}

\item{db}{a dataset containing annotation information. In lucky package, it is often \code{common.annot} dataset for \code{hg38}.}

\item{enhanceList}{A list with Symbol-EnsemblID pairs}
}
\description{
enhance luckyBase::convert with extra annotation
}
\examples{
enhanceList = list(
  'FAM101B' = 'ENSG00000183688',
  'APITD1' = 'ENSG00000175279',
  'C17orf70' = 'ENSG00000185504',
  'SHFM1' = 'ENSG00000127922',
  'MCEMP1' = 'ENSG00000183019',
  'FGF7P3' = 'ENSG00000204837',
  'LOC100506718' = 'ENSG00000185070',
  'MAGI2-AS3' = 'ENSG00000234456',
  'LINC01279' = 'ENSG00000091986',
  'ZNF667-AS1' = 'ENSG00000166770',
  'LOC100507334' = 'ENSG00000261760',
  'ZNF542P' = 'ENSG00000240225',
  'C10orf54' = 'ENSG00000107738',
  "SEPTIN4" = "ENSG00000108387",
  "LINC02381" = "ENSG00000250742",
  "AC005747.1"  = "ENSG00000007237",
  "AC016747.1" = "ENST00000420918",
  "CALHM5" = "ENSG00000178033",
  "CAVIN1" = "ENSG00000177469",
  "CCN4" = "ENSG00000104415",
  "DIPK2B" = "ENSG00000147113",
  "GSDME" = "ENSG00000105928",
  "GUCY1A1" = "ENSG00000164116",
  "GUCY1B1" = "ENSG00000061918",
  "JCAD" = "ENSG00000165757",
  "LHFPL6" = "ENSG00000183722",
  "MAGI2-AS3" = "ENSG00000234456",
  "MSC-AS1" = "ENSG00000235531",
  "NIBAN1" = "ENSG00000135842",
  "PNMA8B" = "ENSG00000204851",
  "RTL5" = "ENSG00000242732",
  "SPART" = "ENSG00000133104",
  "TENM4" = "ENSG00000149256",
  "ZNF667-AS1" = "ENSG00000166770"
)
}
\seealso{
\code{\link{convert}}
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
