% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Fastextra.R
\name{Fastextra}
\alias{Fastextra}
\title{Fastextra}
\usage{
Fastextra(vt, split, n = NULL)
}
\arguments{
\item{vt}{A character vector}

\item{split}{The split string}

\item{n}{The position of string you want to select. If \code{NULL}, then all the substrings would be used.}
}
\value{
a character or a list
}
\description{
A fast way to extra part of elements in a vector
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
