% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Fastgrep.R
\name{Fastgrep}
\alias{Fastgrep}
\title{Fastgrep}
\usage{
Fastgrep(pattern, x)
}
\arguments{
\item{pattern}{pattern string}

\item{x}{a vector}
}
\value{
the position of the pattern-contained components
}
\description{
A plus version of \code{base::grep}
}
\seealso{
\code{\link[base]{grep}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
