% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_flatten_list.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a complex list to a 1-D list}
\usage{
flatten_list(lst)
}
\arguments{
\item{lst}{a list}
}
\value{
1-D list
}
\description{
Flatten a complex list to a 1-D list
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
