% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FPKM2TPM.R
\name{FPKM2TPM}
\alias{FPKM2TPM}
\title{FPKM to TPM}
\usage{
FPKM2TPM(mt, verbose = T)
}
\arguments{
\item{mt}{FPKM matrix (row is gene, column is sample)}

\item{verbose}{Whether to report the process}
}
\value{
TPM matrix
}
\description{
RNA-Seq: FPKM to TPM
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
