% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_importPackage.R
\name{importPackage}
\alias{importPackage}
\title{Fast way to use import for a new package}
\usage{
importPackage(path.namespace = ".")
}
\arguments{
\item{path.namespace}{the path of NAMESPACE file of the new package}
}
\value{
a complete import relationship in DESCRIPTION file of a new package
}
\description{
Fast way to use import for a new package
}
\details{
1. The package must be managed via \code{\link[roxygen2]{roxygen2-package}}; \cr 2. The package import relationship must be updated via \code{\link[roxygen2]{roxygenise}}
}
\note{
before use \code{importPackage()}, you have to run \code{roxygen2::roxygenise()} before in order to gather correct package usage from this new package.
}
\examples{
# If you are in the space of a new package, just
importPackage()
}
\seealso{
\code{\link[roxygen2]{roxygenise}}; \code{\link[devtools]{use_package}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
