% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_is.all.na.R
\name{is.all.na}
\alias{is.all.na}
\title{Is all NA?}
\usage{
is.all.na(vt)
}
\arguments{
\item{vt}{a vector}
}
\value{
logic
}
\description{
Is all NA?
}
\examples{
a <- c(NA,NA);is.all.na(a)
b <- c(NA,2);is.all.na(b)
c <- c(1,3);is.all.na(c)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
