% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_is.all.true.R
\name{is.all.true}
\alias{is.all.true}
\title{Is all True?}
\usage{
is.all.true(vt)
}
\arguments{
\item{vt}{a vector}
}
\value{
logic
}
\description{
Is all True?
}
\examples{
a <- c(T,T,T);is.all.true(a)
a <- c(T,T,F);is.all.true(a)
b <- c(T,2);is.all.true(b)
c <- c(F,3);is.all.true(c)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
