% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_is.one.na.R
\name{is.one.na}
\alias{is.one.na}
\title{Is one NA?}
\usage{
is.one.na(vt)
}
\arguments{
\item{vt}{a vector}
}
\value{
logic
}
\description{
Is one NA?
}
\examples{
a <- c(NA,NA);is.one.na(a)
b <- c(NA,2);is.one.na(b)
c <- c(1,3);is.one.na(c)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
