% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_is.one.true.R
\name{is.one.true}
\alias{is.one.true}
\title{Is one True?}
\usage{
is.one.true(vt)
}
\arguments{
\item{vt}{a vector}
}
\value{
logic
}
\description{
Is one True?
}
\examples{
vt <- c(T,"2","1","3");is.one.true(vt)
vt <- c(F,"2","1","3");is.one.true(vt)
vt <- c(T,2,1,3);is.one.true(vt)
vt <- c(T,F);is.one.true(vt)
vt <- c(F,F);is.one.true(vt)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
