% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_convert.R
\name{Plus.convert}
\alias{Plus.convert}
\title{multiple annotation based on dataset common.annot.}
\usage{
Plus.convert(
  data,
  type,
  fromtype = c("ENTREZID", "SYMBOL"),
  totype = "ENSEMBL",
  goldstandard = "ENTREZID",
  db = common.annot
)
}
\arguments{
\item{data}{a list of feature data}

\item{fromtype}{one or more fromtype, like "ENTREZID","SYMBOL", and "ENSEMBL"}

\item{totype}{one target totype, like "ENTREZID","SYMBOL", and "ENSEMBL"}

\item{goldstandard}{if annotated with different results, select this type as gold standard}

\item{db}{a dataset containing annotation information. In lucky package, it is often \code{common.annot} dataset for \code{hg38}.}
}
\description{
multiple annotation based on dataset common.annot.Its colnames include:"ENSEMBL","ENSEMBL.Versions","SYMBOL","GENENAME","ENTREZID","UNIGENE" and "gene_type".
}
\details{
a plus version of \link{convert} for GEO annotation
}
