% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_rowScale.R
\name{rowScale}
\alias{rowScale}
\title{Row scale for expression matrix}
\usage{
rowScale(matrix, log.convert = T)
}
\arguments{
\item{matrix}{matrix with gene row and sample col}

\item{log.convert}{whether to do log2 convertion before scaling}
}
\description{
Row scale for expression matrix
}
\seealso{
\code{\link[base]{scale}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
