% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_show.color.R
\name{show.color}
\alias{show.color}
\title{Show a plot with colors}
\usage{
show.color(yourcolor)
}
\arguments{
\item{yourcolor}{a string of color. See \code{\link{mycolor}}}
}
\value{
a vector and a ggplot
}
\description{
Show a plot with colors
}
\examples{
show.color(mycolor)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
