% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_similarity.R
\name{similarity}
\alias{similarity}
\title{Caluculate similarity of two vectors}
\usage{
similarity(vec1, vec2, method = "jaccard")
}
\arguments{
\item{vec1}{A vector}

\item{vec2}{A vector}

\item{method}{One of \code{"jaccard"}
\itemize{
  \item \code{jaccard} calculate similarity that doesn't depend on the order of elements in the vectors
}}
}
\value{
Numeric that defines similarity.
}
\description{
Caluculate similarity of two vectors
}
\examples{
vec1 <- c(1, 2, 3, 4, 5, 1, 2)
vec2 <- c(3, 4, 5, 6, 7, 3, 4)
sim <- similarity(vec1, vec2, method='jaccard')
luckyVerbose("Jaccard similarity:", round(sim, 4))
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
