% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_verbose_vector.R
\name{verbose_vector}
\alias{verbose_vector}
\title{verbose_vector}
\usage{
verbose_vector(vector, tablename = T)
}
\arguments{
\item{vector}{a vector}

\item{tablename}{whether verbose table names of the vector}
}
\value{
a character
}
\description{
verbose a vector. For example, report \code{c(1,2,3)} as \code{"1", "2", "3"}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
