% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ReadCount2FPKM.R
\name{ReadCount2FPKM}
\alias{ReadCount2FPKM}
\title{RNA-Seq Read Count to FPKM}
\usage{
ReadCount2FPKM(readcount, geneid = "ensembl", genomic = "hg38")
}
\arguments{
\item{readcount}{Read count matrix(row is gene, column is sample)}

\item{geneid}{Only support \code{ensembl} now}

\item{genomic}{Only support \code{hg38} now}
}
\value{
FPKM matrix
}
\description{
RNA-Seq Read Count to FPKM
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
